fix: merge provider in newtab
